progress on pvalue filter / drag still not working
