+++ Service/src/main/java/org/niagads/genomics/service/ServiceApplication.java
@@ -4,7 +4,7 @@ import java.util.Set;
 
 import org.eupathdb.common.service.EuPathServiceApplication;
 
-import org.niagads.genomics.service.services.Search.SiteSearchService;
+import org.niagads.genomics.service.services.Search.*;
 import org.niagads.genomics.service.services.LocusZoom.*;
 import org.niagads.genomics.service.services.Dataset.*;
 import org.niagads.genomics.service.services.Variant.*;
@@ -28,6 +28,7 @@ public class ServiceApplication extends EuPathServiceApplication {
     .add(LZRecombinationService.class)
 
     .add(SiteSearchService.class)
+    .add(OntologySearchService.class)
 
     .add(DatasetModelRefService.class)
     .add(DatasetLookupService.class)
