+++ Site/webapp/wdkCustomization/js/client/components/RecordPage/RecordTable/RecordTableFilters/index.ts
@@ -1,2 +1,2 @@
 export * from './PieChartFilter';
-export * from './PValueChartFilter';
+export * from './PValueFilter';
